+++ src/matrix-auth/SecurityManager.ts
