+++ src/matrix-auth/auth/util/verifyServerConfig.ts
