+++ src/matrix-auth/unexported/verificationMethods.ts
