+++ src/matrix-yjs/MatrixProvider.ts
@@ -3,7 +3,7 @@ import { createClient, MatrixClient } from "matrix-js-sdk";
 import * as Y from "yjs";
 import { Emitter, Event } from "../util/vscode-common/event";
 import { Disposable } from "../util/vscode-common/lifecycle";
-import { decodeBase64, encodeBase64 } from "./unexported/olmlib";
+import { decodeBase64, encodeBase64 } from "../matrix-auth/unexported/olmlib";
 
 const FLUSH_INTERVAL = 1000 * 5;
